# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Flashcards help you learn more vocabulary anytime, anywhere.",
        "title": "Word Flashcard API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/questions": {
            "get": {
                "description": "Get all questions, supports pagination and multi-column sorting through query parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "questions"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Maximum number of records to return (default: 100, max: 1000)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of records to skip (default: 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort columns/expressions, comma-separated. Format: col,-col,(expr),-(expr). Allowed: id,question,answer,count_practise,count_failure_practise,created_at,updated_at",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Questions retrieved successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/word-flashcard_internal_models.Question"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid query parameters",
                        "schema": {
                            "$ref": "#/definitions/word-flashcard_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error - Failed to fetch data from database",
                        "schema": {
                            "$ref": "#/definitions/word-flashcard_internal_models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new question entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "questions"
                ],
                "parameters": [
                    {
                        "description": "Question data to create",
                        "name": "question",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/word-flashcard_internal_models.Question"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question created successfully",
                        "schema": {
                            "$ref": "#/definitions/word-flashcard_internal_models.Question"
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/word-flashcard_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error - Failed to insert data into database",
                        "schema": {
                            "$ref": "#/definitions/word-flashcard_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/questions/count": {
            "get": {
                "description": "Get the total count of questions in the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "questions"
                ],
                "responses": {
                    "200": {
                        "description": "Total count of questions",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer",
                                "format": "int64"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error - Failed to count questions in database",
                        "schema": {
                            "$ref": "#/definitions/word-flashcard_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/questions/random": {
            "post": {
                "description": "Randomly obtain the required number of questions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "questions"
                ],
                "parameters": [
                    {
                        "description": "Random filter criteria including count and optional filter",
                        "name": "randomFilter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/word-flashcard_internal_models.QuestionRandomRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Random questions retrieved successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/word-flashcard_internal_models.Question"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid request body or count parameter",
                        "schema": {
                            "$ref": "#/definitions/word-flashcard_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error - Failed to fetch data from database",
                        "schema": {
                            "$ref": "#/definitions/word-flashcard_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/questions/stats": {
            "get": {
                "description": "Get question count distribution by accuracy rate in 10% intervals, each bucket further broken down by practice count",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "questions"
                ],
                "responses": {
                    "200": {
                        "description": "Question accuracy distribution",
                        "schema": {
                            "$ref": "#/definitions/word-flashcard_internal_models.QuestionStats"
                        }
                    },
                    "500": {
                        "description": "Internal server error - Failed to fetch questions",
                        "schema": {
                            "$ref": "#/definitions/word-flashcard_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/questions/trend": {
            "get": {
                "description": "Get daily practice count / accuracy rate across all questions over the last N days, zero-filled for days with no activity",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "questions"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of days to include (default: 30, max: 90)",
                        "name": "days",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daily trend points, ascending by date",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/word-flashcard_internal_models.QuestionTrendPoint"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid days parameter",
                        "schema": {
                            "$ref": "#/definitions/word-flashcard_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error - Failed to fetch data from database",
                        "schema": {
                            "$ref": "#/definitions/word-flashcard_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "word-flashcard_internal_models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/word-flashcard_internal_models.ErrorCode"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "word-flashcard_internal_models.Question": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "count_failure_practise": {
                    "type": "integer"
                },
                "count_practise": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "option_a": {
                    "type": "string"
                },
                "option_b": {
                    "type": "string"
                },
                "option_c": {
                    "type": "string"
                },
                "option_d": {
                    "type": "string"
                },
                "practiced": {
                    "type": "boolean"
                },
                "question": {
                    "type": "string"
                },
                "reference": {
                    "type": "string"
                },
                "selected_option": {
                    "type": "string"
                }
            }
        },
...
        "word-flashcard_internal_models.QuestionStats": {
            "type": "object",
            "properties": {
                "accuracy_distribution": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/word-flashcard_internal_models.AccuracyBucket"
                    }
                }
            }
        },
...
    }
}